mock test api added

+++ lib/features/home/model/course_response.dart
@@ -42,7 +42,7 @@ class Course {
   String duration;
   String summary;
   DateTime lastUpdate;
-  int coursePrice;
+  double coursePrice;
   String courseFeatureImage;
   String status;
   DateTime createdAt;
